[modify]

+++ src/main/java/com/xue/repository/dao/UserMapper.java
@@ -598,7 +598,7 @@ public interface UserMapper {
 
     List<Lesson> getLessonById(Integer id);
 
-    Integer getLessonAllCountByDayUnconfirmed(String studio,String campus,String create_time);
+    Float getLessonAllCountByDayUnconfirmed(String studio,String campus,String create_time);
 
     Integer getLessonAllCountByDayByName(String studio,Integer day,String duration,String class_number,String subject,String student_name,String campus);
